+++ tools/run_gdb.sh
@@ -1,1 +1,1 @@
-env PYTHONPATH=$HOME/projects/ppmd-cffi/build/lib.linux-x86_64-3.8/ LD_LIBRARY_PATH==$HOME/projects/ppmd-cffi/build/lib.linux-x86_64-3.8/ gdb tools/pytest_runner
+env PYTHONPATH=$HOME/Projects/ppmd-cffi/build/lib.linux-x86_64-3.8/ LD_LIBRARY_PATH==$HOME/Projects/ppmd-cffi/build/lib.linux-x86_64-3.8/ gdb tools/pytest_runner