+++ Changelog.rst
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 Added
 -----
 
+* Add PPMd8 sources
+
 Changed
 -------